# Improve splitAtBoundary performance