--- a/multipart.cabal
+++ b/multipart.cabal
@@ -29,6 +29,6 @@ library
     Network.Multipart.Header
   build-depends:
       base >= 3 && < 5
-    , bytestring
+    , bytestring >= 0.10.8.0 && < 0.11
     , parsec >= 2.0
     , stringsearch