I replaced the custom substring function with the one from `stringsearch`, which yielded a nice performance bump when using `parseMultipartBody`.

Please let me know if you'd like any changes made before upstreaming.

When parsing an email with 6MB of attachments, I recorded the following performance:
### Before this commit:
```
        Thu Aug  2 08:23 2018 Time and Allocation Profiling Report  (Final)

           main +RTS -p -RTS ./attachments.mbox

        total time  =       33.56 secs   (33563 ticks @ 1000 us, 1 processor)
        total alloc = 444,337,760 bytes  (excludes profiling overheads)

COST CENTRE            MODULE                             SRC                                                   %time %alloc

splitAtBoundary.spl.s2 Network.Multipart                  src/Network/Multipart.hs:132:19-40                     49.7    0.0
splitAtBoundary.spl    Network.Multipart                  src/Network/Multipart.hs:(126,3)-(133,41)              49.3    0.9
lines8                 Text.PortableLines.ByteString.Lazy Text/PortableLines/ByteString/Lazy.hs:(19,1)-(21,48)    0.4   33.3
b64strip               Email.Parse                        src/Email/Parse.hs:252:1-80                             0.3   35.3
findCRorLF             Network.Multipart                  src/Network/Multipart.hs:207:1-60                       0.1    2.4
breakDelim             Data.List.Split.Internals          src/Data/List/Split/Internals.hs:(151,1)-(156,36)       0.0    2.9
mboxMessages.f.\       Email.Parse.MBox                   src/Email/Parse/MBox.hs:33:17-73                        0.0    4.5
satisfy                Text.Parsec.Char                   Text/Parsec/Char.hs:(140,1)-(142,71)                    0.0    1.7
findCRLF               Network.Multipart                  src/Network/Multipart.hs:(197,1)-(204,52)               0.0    2.6
decodeUtf8With         Data.Text.Encoding                 Data/Text/Encoding.hs:(132,1)-(155,71)                  0.0    2.9
encodeUtf8             Data.Text.Encoding                 Data/Text/Encoding.hs:(373,1)-(388,37)                  0.0    5.7
openMbox               Email.Parse.MBox                   src/Email/Parse/MBox.hs:21:1-46                         0.0    1.5
splitAtBoundary.spl.s1 Network.Multipart                  src/Network/Multipart.hs:131:19-38                      0.0    3.3
```

### After this commit:

```
        Thu Aug  2 08:24 2018 Time and Allocation Profiling Report  (Final)

           main +RTS -p -RTS ./attachments.mbox

        total time  =        0.25 secs   (250 ticks @ 1000 us, 1 processor)
        total alloc = 564,201,000 bytes  (excludes profiling overheads)

COST CENTRE           MODULE                             SRC                                                   %time %alloc

b64strip              Email.Parse                        src/Email/Parse.hs:252:1-80                            37.2   27.8
lines8                Text.PortableLines.ByteString.Lazy Text/PortableLines/ByteString/Lazy.hs:(19,1)-(21,48)   31.6   26.2
splitAtBoundary.(...) Network.Multipart                  src/Network/Multipart.hs:125:7-49                      18.0   28.5
breakDelim            Data.List.Split.Internals          src/Data/List/Split/Internals.hs:(151,1)-(156,36)       2.4    2.3
satisfy               Text.Parsec.Char                   Text/Parsec/Char.hs:(140,1)-(142,71)                    2.0    1.3
mboxMessages.f.\      Email.Parse.MBox                   src/Email/Parse/MBox.hs:33:17-73                        1.6    3.6
main.\.\.\            Main                               src-bin/main.hs:20:20-113                               1.2    0.0
openMbox              Email.Parse.MBox                   src/Email/Parse/MBox.hs:21:1-46                         1.2    1.2
decodeUtf8With        Data.Text.Encoding                 Data/Text/Encoding.hs:(132,1)-(155,71)                  0.8    2.3
encodeUtf8            Data.Text.Encoding                 Data/Text/Encoding.hs:(373,1)-(388,37)                  0.4    4.5
```