@hesselink 2018-08-15
Thanks! I've released this as 0.1.3.